talking about metadata
Metadata can be complicated. At some point I should figure out how
track-specific metadata interacts with non-track specific metadata
(both declared). Because metadata is stored in a hash table, there's
not exactly an easy way to tell what order it will come out in.

I would prefer not to make a performance hit in order to ensure
track-specific metadata always overrides non-track-specific
metadata. If I can figure out a good way to cache it all, that would
be dandy. It's a bit expensive as it is.