improved metadata handling
Move parsing of metadata into declare-time. This should speed the
command building loop greatly on longer albums or when given lots of
metadata... and especially both.

When building the command, if both the track-specific metadata and
non-track-specific metadata are not declared, skip over them.

When track specific metadata is given, it will always override
non-track-specific metadata.

If track-specific metadata is given for artist or album, these are not
used to compute save path.